data load v2

+++ src/main/java/com/evoltech/ciqapm/model/Servicio.java
@@ -24,6 +24,6 @@ public class Servicio extends BaseClass {
 
     private String entregableEsperado;
 
-    private BigDecimal costo;
+    private Double costo;
 
 }
